requirements.txt
asgiref==3.4.1
brotlipy==0.7.0
certifi==2021.5.30
cffi
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography 
Django==3.2
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna 
importlib-metadata==4.8.1
itypes==1.2.0
Jinja2 
Markdown==3.3.4
MarkupSafe 
pycparser
PyJWT==2.1.0
pyOpenSSL 
PySocks 
pytz==2021.3
PyYAML==5.4.1
requests 
six
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
win-inet-pton 
wincertstore==0.2
zipp==3.6.0